* Added support for providing `key=value` command line parameters instead of only positional. * The parse_args() function will now also return base_args (and base_kwargs), which can be used to configure the app behaviour before starting the action chain. * Loading config in action_cli, so we can use a permanent config, e.g. to change `api_url` to use the v7.1 Sync API url. * Added new convenience actions, `-due` (alias for `-is due`), and `-not` (alias for `-is not`), and `-name` (alias for `-content_iglob_filter`).
I've started preparing to upgrade to v7.1 Sync API, because I realized there was no way to reliably determine if a task is recurring in the v7.0 API.
Unfortunately, supporting both versions adds a bit of complexity.
Specifically, the v7.1 API deals with due dates differently. They are saved as a dict attribute, rather than just a simple string,
The format has also changed to ISO8601, and instead of using "23:59:59" as "all day" due time, it just sets due.date to 'YYYY-MM-DD' with no time spec.